pyproject.toml
[project]
name = "shotgraph"
version = "0.1.0"
description = "AI Cinematic Video Generation Pipeline"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "ShotGraph Team"}
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "moviepy>=1.0.3",
    "pillow>=10.2.0",
    "toons>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]
gpu = [
    "torch>=2.1.0",
    "diffusers>=0.25.0",
    "transformers>=4.37.0",
    "accelerate>=0.26.0",
    "xformers>=0.0.23",
    "scipy>=1.11.0",
    "edge-tts>=6.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
